+++ Helpers/Missions/Assassinate/PerformAssassinationMissionStep.cs
@@ -23,7 +23,7 @@ namespace OnlyWar.Helpers.Missions.Assassinate
             // size 2: Broodlord
             // size 3: Hive Tyrant
             RegionFaction enemyFaction = context.Order.Mission.RegionFaction;
-            float difficulty = (enemyFaction.Entrenchment + enemyFaction.GetOwnRegionIntel()) * 0.5f
+            float difficulty = (float)((enemyFaction.Entrenchment + enemyFaction.GetOwnRegionIntel()) * 0.5)
                 + (float)Math.Log10(enemyFaction.Garrison);
             LeaderMissionTest missionTest = new LeaderMissionTest(tactics, difficulty);
             float margin = missionTest.RunMissionCheck(context.MissionSquads);
