+++ Helpers/Missions/Sabotage/PerformSabotageMissionStep.cs
@@ -17,7 +17,7 @@ namespace OnlyWar.Helpers.Missions.Sabotage
         {
             BaseSkill tactics = GameDataSingleton.Instance.GameRulesData.Skills.Tactics;
             RegionFaction enemyFaction = context.Order.Mission.RegionFaction;
-            float difficulty = enemyFaction.Entrenchment * 0.5f;
+            float difficulty = (float)(enemyFaction.Entrenchment * 0.5);
             difficulty += (float)Math.Log10(enemyFaction.Garrison);
             LeaderMissionTest missionTest = new LeaderMissionTest(tactics, difficulty);
